Refactor font handling into dedicated modules
