+++ src/client/client.hpp
@@ -1286,6 +1286,9 @@ extern cl_scr_t scr;
 
 void    SCR_Init(void);
 void    SCR_Shutdown(void);
+void    SCR_InitFontSystem(void);
+void    SCR_ShutdownFontSystem(void);
+void    SCR_RefreshFontCvar(void);
 void    SCR_UpdateScreen(void);
 void    SCR_SizeUp(void);
 void    SCR_SizeDown(void);
